Add common component for idp mapping method dropdown
